Refactoring CameraWrapper
