--- app/src/main/java/com/github/orkest/ui/Camera/CameraWrapper.kt
@@ -1,44 +1,0 @@
-package com.github.orkest.ui.Camera
-
-import android.content.Context
-import android.os.Build
-import androidx.annotation.RequiresApi
-import androidx.camera.core.CameraSelector
-import androidx.camera.core.Preview
-import androidx.camera.lifecycle.ProcessCameraProvider
-import androidx.camera.video.*
-import androidx.camera.view.PreviewView
-import androidx.core.util.Consumer
-import androidx.lifecycle.LifecycleOwner
-import java.io.File
-import java.text.SimpleDateFormat
-import java.util.*
-import java.util.concurrent.Executor
-import kotlin.coroutines.resume
-import kotlin.coroutines.suspendCoroutine
-
-
-fun startRecordingVideo(
-    context: Context,
-    filenameFormat: String,
-    videoCapture: VideoCapture<Recorder>,
-    outputDirectory: File,
-    executor: Executor,
-    audioEnabled: Boolean,
-    consumer: Consumer<VideoRecordEvent>
-): Recording {
-    val videoFile = File(
-        outputDirectory,
-        SimpleDateFormat(filenameFormat, Locale.US).format(System.currentTimeMillis()) + ".mp4"
-    )
-
-    val outputOptions = FileOutputOptions.Builder(videoFile).build()
-
-    //Start the recording process with the given executor and consumer
-    return videoCapture.output
-        .prepareRecording(context, outputOptions)
-        .start(executor, consumer)
-}
-
-
-
